fix for incorrect loading
