+++ src/bot/api.ts
@@ -8,7 +8,6 @@ import { Socket } from 'socket.io';
 import { getManager, getRepository, In, IsNull, Not } from 'typeorm';
 
 import Core from './_interface';
-import { getBroadcaster, isBot, isBroadcaster, isIgnored } from './commons';
 import * as constants from './constants';
 import customvariables from './customvariables';
 import { CacheGames } from './database/entity/cacheGames';
@@ -20,7 +19,11 @@ import { getFunctionList, onStartup } from './decorators/on';
 import events from './events';
 import { isDbConnected } from './helpers/database';
 import { dayjs } from './helpers/dayjs';
+import { getBroadcaster } from './helpers/getBroadcaster';
 import { triggerInterfaceOnFollow } from './helpers/interface/triggers';
+import { isBot } from './helpers/isBot';
+import { isBroadcaster } from './helpers/isBroadcaster';
+import { isIgnored } from './helpers/isIgnored';
 import { debug, error, follow, info, start, stop, unfollow, warning } from './helpers/log';
 import { ioServer } from './helpers/panel';
 import { linesParsed, setStatus } from './helpers/parser';
