+++ src/bot/helpers/getBroadcaster.ts
@@ -1,0 +1,9 @@
+import oauth from '../oauth';
+
+export function getBroadcaster() {
+  try {
+    return oauth.broadcasterUsername.toLowerCase().trim();
+  } catch (e) {
+    return '';
+  }
+}
