+++ src/bot/helpers/isModerator.ts
@@ -1,0 +1,8 @@
+import type { UserInterface } from '../database/entity/user';
+
+export function isModerator(user: UserInterface | UserStateTags): boolean {
+  if ('mod' in user) {
+    return user.mod === '1';
+  }
+  return user.isModerator ?? false;
+}
