+++ src/bot/integrations/streamlabs.ts
@@ -5,12 +5,12 @@ import { io, Socket } from 'socket.io-client';
 import { getRepository } from 'typeorm';
 
 import api from '../api';
-import { getBroadcaster } from '../commons';
 import currency from '../currency';
 import { User, UserTipInterface } from '../database/entity/user';
 import { persistent, settings, ui } from '../decorators';
 import { onChange, onStartup } from '../decorators/on';
 import events from '../events';
+import { getBroadcaster } from '../helpers/getBroadcaster';
 import { triggerInterfaceOnTip } from '../helpers/interface/triggers';
 import { debug, error, info, tip } from '../helpers/log';
 import { ioServer } from '../helpers/panel';
