+++ src/bot/systems/songs.ts
@@ -7,10 +7,11 @@ import ytdl from 'ytdl-core';
 import ytpl from 'ytpl';
 import ytsr from 'ytsr';
 
-import { announce, getBot, getBotSender, isModerator, prepare, timeout } from '../commons';
+import { announce, getBot, getBotSender, prepare, timeout } from '../commons';
 import { SongBan, SongPlaylist, SongPlaylistInterface, SongRequest } from '../database/entity/song';
 import { command, default_permission, persistent, settings, ui } from '../decorators';
 import { onChange, onStartup } from '../decorators/on';
+import { isModerator } from '../helpers/isModerator';
 import { error, info } from '../helpers/log';
 import { permission } from '../helpers/permissions';
 import { adminEndpoint, publicEndpoint } from '../helpers/socket';
